requirement.txt
networkx==2.8.4
nltk==3.7
numpy==1.23.5
pdfminer.six==20191110
scikit_learn==1.2.1
